feat: 나머지 화면 및 viewModel 완료

+++ app/src/main/java/com/example/movieismylife/model/CommentView.kt
@@ -1,7 +1,7 @@
 package com.example.movieismylife.model
 
 data class CommentView (
-    val score: Float = 0f,
+    val score: Long = 0,
     val content: String = "",
     val name: String = "",
     val profile: String = "",
